Fixed breaking bug after upgrading Realm.

+++ Source/Managers/Data Manager/DataManager.m
@@ -48,7 +48,7 @@ static NSString * const kLatestComicDownloadedKey = @"LatestComicDownloaded";
 
 - (void)initializeRealm {
     // Currently no need for migrations, so we can leave the migrations block empty.
-    [RLMRealm setSchemaVersion:kCurrentSchemaVersion forRealmAtPath:[RLMRealm defaultRealmPath] withMigrationBlock:^(RLMMigration *migration, NSUInteger oldSchemaVersion) {}];
+    [RLMRealm setSchemaVersion:kCurrentSchemaVersion forRealmAtPath:[RLMRealm defaultRealmPath] withMigrationBlock:^(RLMMigration *migration, uint64_t oldSchemaVersion) {}];
 
     self.realm = [RLMRealm defaultRealm];
 }
